+++ NEWS.md
@@ -1,6 +1,7 @@
 # ruminate 0.2.0 
 - Added MB (model builder) module with
   - Support for rxode2 and NONMEM models
+- Fixed bug in NCA module where column details didn't update properly when switching between analyses 
 
 # ruminate 0.1.1
 - Initial release